teste-1/teste1/requirements-dev.txt
astroid==2.3.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
djangorestframework==3.10.3
django-rest-swagger==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.2
pylint==2.4.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
